swc_ecma_lexer::common::parser

Module expr

Functions

at_possible_async

finish_assignment_expr

parse_args

Parse Arguments[Yield, Await]

parse_args_or_pats

parse_array_lit

parse_assignment_expr

parseMaybeAssign (overridden)

parse_await_expr

parse_bin_expr

Name from spec: ‘LogicalORExpression’

parse_bin_op_recursively

Parse binary operators with the operator precedence parsing algorithm. left is the left-hand side of the operator. minPrec provides context that allows the function to stop and defer further parser to one of its callers when it encounters an operator that has a lower precedence than the set it is parsing.

parse_dynamic_import_or_import_meta

parse_lhs_expr

Parse call, dot, and []-subscript expressions.

parse_lit

parse_member_expr_or_new_expr

is_new_expr: true iff we are parsing production ‘NewExpression’.

parse_new_expr

Parse NewExpression. This includes MemberExpression.

parse_paren_expr_or_arrow_fn

parse_primary_expr_rest

parse_str_lit

parse_subscripts

parse_this_expr

try_parse_async_start

try_parse_regexp