swc_ecma_lexer::common::parser

Module typescript

Enums

ParsingContext

Functions

eat_any_ts_modifier

is_ts_start_of_mapped_type

tsIsStartOfMappedType

parse_ts_entity_name

tsParseEntityName

parse_ts_enum_decl

tsParseEnumDeclaration

parse_ts_expr_stmt

tsParseExpressionStatement

parse_ts_heritage_clause

tsParseHeritageClause

parse_ts_import_equals_decl

tsParseImportEqualsDeclaration

parse_ts_interface_decl

tsParseInterfaceDeclaration

parse_ts_mapped_type

tsParseMappedType

parse_ts_modifier

Parses a modifier matching one the given modifier names.

parse_ts_parenthesized_type

tsParseParenthesizedType

parse_ts_this_type_node

tsParseThisTypeNode

parse_ts_this_type_predicate

tsParseThisTypePredicate

parse_ts_tuple_type

tsParseTupleType

parse_ts_type

Be sure to be in a type context before calling p.

parse_ts_type_alias_decl

tsParseTypeAliasDeclaration

parse_ts_type_ann

parse_ts_type_args

tsParseTypeArguments

parse_ts_type_assertion

tsParseTypeAssertion

parse_ts_type_lit

tsParseTypeLiteral

parse_ts_type_or_type_predicate_ann

tsParseTypeOrTypePredicateAnnotation

parse_ts_type_params

tsParseTypeParameter

parse_ts_type_ref

tsParseTypeReference

try_parse_ts

tsTryParse

try_parse_ts_declare

tsTryParseDeclare

try_parse_ts_export_decl

tsTryParseExportDeclaration

try_parse_ts_generic_async_arrow_fn

tsTryParseGenericAsyncArrowFunction

try_parse_ts_index_signature

tsTryParseIndexSignature

try_parse_ts_type_ann

tsTryParseTypeAnnotation

try_parse_ts_type_or_type_predicate_ann

tsTryParseTypeOrTypePredicateAnnotation

try_parse_ts_type_params

tsTryParseTypeParameters

ts_look_ahead